Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    float a,b;
+    int a,b;
     char c;
-    scanf("%f %f",&a,&b);
+    scanf("%i %i",&a,&b);
     scanf(" %c",&c);
     int x = a+b;
     int y = a-b;
